Added optional folder cleanup

+++ test/Chain/test_GippFile.py
@@ -81,8 +81,9 @@ class TestGippFile(unittest.TestCase):
         n_extl = len(FileSystem.find("*EEF", g.out_path))
         self.assertEqual(n_extl, 15)
         FileSystem.remove_file(os.path.join(self.root, "wget-log"))
-        FileSystem.remove_directory(g.out_path)
-        self.assertFalse(os.path.exists(g.out_path))
+        if not os.getcwd() == g.out_path:
+            FileSystem.remove_directory(g.out_path)
+            self.assertFalse(os.path.exists(g.out_path))
 
     def test_download_s2_muscate_nocams(self):
         from Common import FileSystem
@@ -99,8 +100,9 @@ class TestGippFile(unittest.TestCase):
         n_extl = len(FileSystem.find("*EEF", g.out_path))
         self.assertEqual(n_extl, 9)
         FileSystem.remove_file(os.path.join(self.root, "wget-log"))
-        FileSystem.remove_directory(g.out_path)
-        self.assertFalse(os.path.exists(g.out_path))
+        if not os.getcwd() == g.out_path:
+            FileSystem.remove_directory(g.out_path)
+            self.assertFalse(os.path.exists(g.out_path))
 
     def test_download_s2_natif_nocams(self):
         from Common import FileSystem
@@ -117,8 +119,9 @@ class TestGippFile(unittest.TestCase):
         n_extl = len(FileSystem.find("*EEF", g.out_path))
         self.assertEqual(n_extl, 9)
         FileSystem.remove_file(os.path.join(self.root, "wget-log"))
-        FileSystem.remove_directory(g.out_path)
-        self.assertFalse(os.path.exists(g.out_path))
+        if not os.getcwd() == g.out_path:
+            FileSystem.remove_directory(g.out_path)
+            self.assertFalse(os.path.exists(g.out_path))
 
     def test_download_l8_muscate_nocams(self):
         from Common import FileSystem
@@ -135,9 +138,10 @@ class TestGippFile(unittest.TestCase):
         n_extl = len(FileSystem.find("*EEF", g.out_path))
         self.assertEqual(n_extl, 5)
         FileSystem.remove_file(os.path.join(self.root, "wget-log"))
-        FileSystem.remove_directory(g.out_path)
-        self.assertFalse(os.path.exists(g.out_path))
-    
+        if not os.getcwd() == g.out_path:
+            FileSystem.remove_directory(g.out_path)
+            self.assertFalse(os.path.exists(g.out_path))
+
     def test_download_vs_muscate_nocams(self):
         from Common import FileSystem
         g = GippSet(self.root, "venus", "muscate")
@@ -153,8 +157,9 @@ class TestGippFile(unittest.TestCase):
         n_extl = len(FileSystem.find("*EEF", g.out_path))
         self.assertEqual(n_extl, 5)
         FileSystem.remove_file(os.path.join(self.root, "wget-log"))
-        FileSystem.remove_directory(g.out_path)
-        self.assertFalse(os.path.exists(g.out_path))
+        if not os.getcwd() == g.out_path:
+            FileSystem.remove_directory(g.out_path)
+            self.assertFalse(os.path.exists(g.out_path))
 
     def test_download_l8_natif_nocams(self):
         from Common import FileSystem
@@ -171,8 +176,9 @@ class TestGippFile(unittest.TestCase):
         n_extl = len(FileSystem.find("*EEF", g.out_path))
         self.assertEqual(n_extl, 5)
         FileSystem.remove_file(os.path.join(self.root, "wget-log"))
-        FileSystem.remove_directory(g.out_path)
-        self.assertFalse(os.path.exists(g.out_path))
+        if not os.getcwd() == g.out_path:
+            FileSystem.remove_directory(g.out_path)
+            self.assertFalse(os.path.exists(g.out_path))
 
     def test_download_vs_natif_nocams(self):
         from Common import FileSystem
@@ -189,8 +195,9 @@ class TestGippFile(unittest.TestCase):
         n_extl = len(FileSystem.find("*EEF", g.out_path))
         self.assertEqual(n_extl, 5)
         FileSystem.remove_file(os.path.join(self.root, "wget-log"))
-        FileSystem.remove_directory(g.out_path)
-        self.assertFalse(os.path.exists(g.out_path))
+        if not os.getcwd() == g.out_path:
+            FileSystem.remove_directory(g.out_path)
+            self.assertFalse(os.path.exists(g.out_path))
 
     def test_symlink(self):
         from Common import FileSystem
@@ -207,6 +214,9 @@ class TestGippFile(unittest.TestCase):
         FileSystem.remove_directory(symlink_dir)
         self.assertFalse(os.path.isdir(symlink_dir))
         FileSystem.remove_file(os.path.join(self.root, "wget-log"))
+        if not os.getcwd() == g.out_path:
+            FileSystem.remove_directory(g.out_path)
+            self.assertFalse(os.path.exists(g.out_path))
 
     def test_get_models_nocams(self):
         from Common import FileSystem
@@ -219,8 +229,9 @@ class TestGippFile(unittest.TestCase):
         models_expected = ["continen"]
         self.assertEqual(g.get_models(), models_expected)
         FileSystem.remove_file(os.path.join(self.root, "wget-log"))
-        FileSystem.remove_directory(g.out_path)
-        self.assertFalse(os.path.exists(g.out_path))
+        if not os.getcwd() == g.out_path:
+            FileSystem.remove_directory(g.out_path)
+            self.assertFalse(os.path.exists(g.out_path))
 
     def test_get_models_cams(self):
         from Common import FileSystem
@@ -233,8 +244,9 @@ class TestGippFile(unittest.TestCase):
         models_expected = sorted(["continen", "dust", "blackcar", "sulphate", "seasalt", "organicm"])
         self.assertEqual(g.get_models(), models_expected)
         FileSystem.remove_file(os.path.join(self.root, "wget-log"))
-        FileSystem.remove_directory(g.out_path)
-        self.assertFalse(os.path.exists(g.out_path))
+        if not os.getcwd() == g.out_path:
+            FileSystem.remove_directory(g.out_path)
+            self.assertFalse(os.path.exists(g.out_path))
 
 
 if __name__ == '__main__':
